refactor(console): move QuickTimeRangeComponent close to where it is used
