+++ gravitee-apim-console-webui/src/components/alerts/quick-time-range/_quick-time-range.scss
