+++ gravitee-apim-console-webui/src/components/alerts/quick-time-range/quick-time-range.component.ts
