+++ gravitee-apim-console-webui/src/index.scss
@@ -34,7 +34,7 @@
   @import 'components/identityPicture/identityPicture';
   @import 'components/image/image';
   @import 'components/documentation/error/error';
-  @import 'components/quick-time-range/quick-time-range';
+  @import 'components/alerts/quick-time-range/quick-time-range';
   @import 'components/user-autocomplete/user-autocomplete';
   @import 'components/widget/widget';
   @import 'components/alerts/dashboard/alerts-dashboard';
